Revert "Add caja extension"
This reverts commit 34162a3ea31b9684a24fa115b82cabacfb8e7eec.

To properly support multiple file renaming caja needs code changes. Using an
action results in duplicated "Rename..." menu entries, the default one, which is
greyed out for multiple selections, native to Caja, and the newly added one via extension
which only appears for multiple selection.

This is messy.

The proper solution is for Caja to either detect bulky, to use a hook, or to have a configuration
option to point to a multiple file renaming tool.