file picker: Don't lose a multi-selection when enter is pressed.
When hitting enter to accept the file selection, it will end up
choosing the last-selected only, even if multiple files are
selected. If a folder is that last-selected item, the picker
will enter the folder instead, which made no sense.

This overrides enter-key handling, and completes the dialog if
multiple items are selected, without allowing the default handler
to change the selection on us.