Never treat folders as though they have extensions.
This renames (and changes the behavior of) the "Name and extension"
scope to "Full name".

Instead of this option modifying the name and extension as separate
strings, it will now work with the full file name.

This scope will be activated and enforced whenever at least one
folder is part of the file list.

Small tweak: Disable the "Rename" button if there are no actual
changes pending.

Fixes #12