Fix issue #34
The program enters infinite loop when the user adds a backslash at the end of a string in "replace" field. It happens because reg.sub throws an exception when it sees a backslash at the end of "replace" variable. The fix was already implemented for the case where regular expression replacement is enabled - catch regex exception and return unchanged string.